# openapi.json
{"openapi":"3.1.0","info":{"title":"Housing Price Prediction API","description":"API for predicting housing prices based on various features","version":"1.0.0"},"paths":{"/predictions":{"post":{"tags":["predictions"],"summary":"Submit a prediction request","description":"Submit a new housing price prediction request","operationId":"submit_prediction_predictions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionSubmissionResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predictions/{run_id}":{"get":{"tags":["predictions"],"summary":"Get prediction result","description":"Get the result of a prediction request by its Dagster run ID","operationId":"get_prediction_predictions__run_id__get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/PredictionPendingResponse"},{"$ref":"#/components/schemas/PredictionCompletedResponse"},{"$ref":"#/components/schemas/PredictionFailedResponse"}],"title":"Response Get Prediction Predictions  Run Id  Get"}}}},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Internal Server Error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ErrorResponse":{"properties":{"detail":{"type":"string","title":"Detail","description":"Error message"}},"type":"object","required":["detail"],"title":"ErrorResponse","description":"Error response model.","example":{"detail":"Invalid input data"}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PredictionCompletedResponse":{"properties":{"run_id":{"type":"string","title":"Run Id","description":"Dagster run ID for the prediction"},"status":{"$ref":"#/components/schemas/PredictionStatus","description":"Status of the prediction"},"prediction":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Prediction","description":"The prediction value if available"},"completed_at":{"type":"string","format":"date-time","title":"Completed At","description":"When the prediction was completed"}},"type":"object","required":["run_id","status","completed_at"],"title":"PredictionCompletedResponse","description":"Response model for completed prediction requests.","example":{"completed_at":"2024-04-10T12:00:05Z","prediction_value":320201.58,"run_id":"123e4567-e89b-12d3-a456-426614174000","status":"completed"}},"PredictionFailedResponse":{"properties":{"run_id":{"type":"string","title":"Run Id","description":"Dagster run ID for the prediction"},"status":{"$ref":"#/components/schemas/PredictionStatus","description":"Status of the prediction"},"prediction":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Prediction","description":"The prediction value if available"},"completed_at":{"type":"string","format":"date-time","title":"Completed At","description":"When the prediction failed"}},"type":"object","required":["run_id","status","completed_at"],"title":"PredictionFailedResponse","description":"Response model for failed prediction requests.","example":{"completed_at":"2024-04-10T12:00:05Z","error":"Error processing prediction","run_id":"123e4567-e89b-12d3-a456-426614174000","status":"failed"}},"PredictionPendingResponse":{"properties":{"run_id":{"type":"string","title":"Run Id","description":"Dagster run ID for the prediction"},"status":{"$ref":"#/components/schemas/PredictionStatus","description":"Status of the prediction"},"prediction":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Prediction","description":"The prediction value if available"}},"type":"object","required":["run_id","status"],"title":"PredictionPendingResponse","description":"Response model for pending/running prediction requests.","example":{"run_id":"123e4567-e89b-12d3-a456-426614174000","status":"pending"}},"PredictionRequest":{"properties":{"longitude":{"type":"number","title":"Longitude","description":"Longitude of the house location"},"latitude":{"type":"number","title":"Latitude","description":"Latitude of the house location"},"housing_median_age":{"type":"number","title":"Housing Median Age","description":"Median age of houses in the block"},"total_rooms":{"type":"number","title":"Total Rooms","description":"Total number of rooms in the block"},"total_bedrooms":{"type":"number","title":"Total Bedrooms","description":"Total number of bedrooms in the block"},"population":{"type":"number","title":"Population","description":"Total population in the block"},"households":{"type":"number","title":"Households","description":"Total number of households in the block"},"median_income":{"type":"number","title":"Median Income","description":"Median income of households in the block"},"ocean_proximity":{"type":"string","enum":["<1H OCEAN","INLAND","ISLAND","NEAR BAY","NEAR OCEAN"],"title":"Ocean Proximity","description":"Proximity to the ocean"}},"type":"object","required":["longitude","latitude","housing_median_age","total_rooms","total_bedrooms","population","households","median_income","ocean_proximity"],"title":"PredictionRequest","description":"Request model for house price prediction.","example":{"households":249.0,"housing_median_age":36.0,"latitude":38.01,"longitude":-122.64,"median_income":5.5789,"ocean_proximity":"NEAR OCEAN","population":678.0,"total_bedrooms":258.0,"total_rooms":1336.0}},"PredictionStatus":{"type":"string","enum":["pending","running","completed","failed","not_found"],"title":"PredictionStatus","description":"Status of a prediction request."},"PredictionSubmissionResponse":{"properties":{"run_id":{"type":"string","title":"Run Id","description":"Dagster run ID for the prediction"},"status":{"$ref":"#/components/schemas/PredictionStatus","description":"Status of the prediction"},"prediction":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Prediction","description":"The prediction value if available"}},"type":"object","required":["run_id","status"],"title":"PredictionSubmissionResponse","description":"Response model for prediction request submission.","example":{"run_id":"123e4567-e89b-12d3-a456-426614174000","status":"pending"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}